Update README.md
The Number Guessing Game is a beginner-friendly project developed using Python and the Tkinter GUI library. This game challenges players to guess a randomly generated number between 1 and 100. It’s an engaging and interactive way to understand the fundamentals of event-driven programming, GUI design, and basic logic implementation in Python.